# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yunq API",
    "description": "Static analysis platform: REST API, OAuth, signed webhooks and operational metrics.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.1"
  },
  "paths": {
...
    "/api/issues": {
      "get": {
        "summary": "Search issues with filters and pagination (newest first).",
        "operationId": "list_issues",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number (default 1).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size (default 50, capped at 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "Filter: info|minor|major|critical|blocker.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter: open|confirmed|resolved|closed.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rule",
            "in": "query",
            "description": "Filter: exact rule id, e.g. owasp:eval-usage.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file",
            "in": "query",
            "description": "Filter: substring of the file path.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "description": "Filter: exact assignee.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facets",
            "in": "query",
            "description": "Comma-separated facets to compute alongside the page:\nseverity, status, rule (e.g. `facets=severity,rule`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching issues",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter value"
          },
          "502": {
            "description": "Storage backend unavailable"
          }
        }
      }
    },
...
    "/api/projects/{key}/blame": {
      "post": {
        "summary": "Ingest per-line SCM blame (the CLI's `--blame-output` JSON, unmodified)\nagainst a project's most recent analysis.",
        "operationId": "ingest_blame",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Branch of the analysis the blame data attaches to (default \"main\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The CLI's `--blame-output` JSON: file path -> ordered per-line blame",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlameLineInput"
                  }
                },
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Blame ingested and persisted"
          },
          "404": {
            "description": "No analysis exists yet for this project/branch — run a scan first"
          },
          "502": {
            "description": "Storage backend failure"
          }
        }
      }
    },
...
    "/api/projects/{key}/sources": {
      "get": {
        "summary": "Per-line issue and coverage annotations for one file; see module docs\nfor what's deliberately out of scope (source text, duplication, blame).",
        "operationId": "sources",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file",
            "in": "query",
            "description": "Exact file path to annotate (as recorded on `issues.file` / ingested\ncoverage reports).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Branch coverage is scoped to (default \"main\"). Issues are not\nbranch-scoped in the current schema — see module docs.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-line issue/coverage annotations for the file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesDto"
                }
              }
            }
          },
          "502": {
            "description": "Storage backend failure"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlameLineInput": {
        "type": "object",
        "description": "One line's blame as the CLI's `--blame-output` JSON encodes it — field\nnames match `blame::BlameLine` exactly so that file can be POSTed as-is.",
        "required": [
          "line",
          "commit",
          "author",
          "author_mail",
          "author_time",
          "summary"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "author_mail": {
            "type": "string"
          },
          "author_time": {
            "type": "integer",
            "format": "int64"
          },
          "commit": {
            "type": "string"
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "FacetCountDto": {
        "type": "object",
        "required": [
          "value",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "value": {
            "type": "string"
          }
        }
      },
      "FacetsDto": {
        "type": "object",
        "required": [
          "by_severity",
          "by_status",
          "by_rule"
        ],
        "properties": {
          "by_rule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetCountDto"
            }
          },
          "by_severity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetCountDto"
            }
          },
          "by_status": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetCountDto"
            }
          }
        }
      },
...
      "ImpactDto": {
        "type": "object",
        "description": "One MQR software-quality impact, serialized alongside the classic type.",
        "required": [
          "quality",
          "severity"
        ],
        "properties": {
          "quality": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          }
        }
      },
      "IssueDto": {
        "type": "object",
        "required": [
          "id",
          "rule",
          "type",
          "impacts",
          "severity",
          "file",
          "line",
          "column",
          "message",
          "status"
        ],
        "properties": {
          "assignee": {
            "type": [
              "string",
              "null"
            ]
          },
          "column": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "file": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "impacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImpactDto"
            }
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "resolution": {
            "type": [
              "string",
              "null"
            ]
          },
          "rule": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "IssuePageDto": {
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "properties": {
          "facets": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FacetsDto",
                "description": "Present only when the request included a `facets` param."
              }
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueDto"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "SourceBlameDto": {
        "type": "object",
        "required": [
          "commit",
          "author",
          "author_mail",
          "author_time",
          "summary"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "author_mail": {
            "type": "string"
          },
          "author_time": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp (seconds) from the commit's `author-time`."
          },
          "commit": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      },
      "SourceIssueDto": {
        "type": "object",
        "required": [
          "id",
          "rule",
          "severity",
          "message"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "rule": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          }
        }
      },
      "SourceLineDto": {
        "type": "object",
        "required": [
          "line",
          "issues"
        ],
        "properties": {
          "blame": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SourceBlameDto",
                "description": "Who last touched this line, from the most recently ingested\n`--blame-output`; absent when no blame has ever been ingested for\nthis file/branch."
              }
            ]
          },
          "coverage_hits": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Times this line executed in the most recently ingested coverage\nreport for this file/branch; absent when that line carries no\ncoverage instrumentation data (or no report has been ingested at\nall — see `coverage_available`).",
            "minimum": 0
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceIssueDto"
            }
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "description": "1-based line number.",
            "minimum": 0
          }
        }
      },
      "SourcesDto": {
        "type": "object",
        "required": [
          "file",
          "coverage_available",
          "lines"
        ],
        "properties": {
          "coverage_available": {
            "type": "boolean",
            "description": "True once a coverage report has ever been ingested that instrumented\nthis file — distinguishes \"measured, zero hits anywhere\" (every\n`coverage_hits` is `Some(0)`) from \"never measured\" (every line's\n`coverage_hits` is absent because this is `false`)."
          },
          "file": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceLineDto"
            },
            "description": "Only lines with at least one issue, coverage data point or blame\nentry are present — this is an annotation overlay, not the full file\n(see module docs on why source text isn't returned)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}